refactor(services): move managers out of utils
Move LanguageManager, ThemeManager, and WindowManager from Utils into Services and update all references. Keep UrlHelper test-friendly by using a fully qualified LanguageManager reference only in the non-test branch.
