+++ CefFlashBrowser/ViewModels/BrowserWindowViewModel.cs
@@ -1,5 +1,6 @@
-﻿using CefFlashBrowser.Data;
+using CefFlashBrowser.Data;
 using CefFlashBrowser.Models;
+using CefFlashBrowser.Services;
 using CefFlashBrowser.Utils;
 using CefSharp;
 using IWshRuntimeLibrary;
