+++ CefFlashBrowser/ViewModels/LanguageSelectorViewModel.cs
@@ -1,4 +1,4 @@
-﻿using CefFlashBrowser.Utils;
+using CefFlashBrowser.Services;
 using SimpleMvvm;
 using SimpleMvvm.Command;
 using System.Collections.Generic;
