+++ CefFlashBrowser/ViewModels/SolNodeViewModel.cs
@@ -1,4 +1,5 @@
-﻿using CefFlashBrowser.Models;
+using CefFlashBrowser.Models;
+using CefFlashBrowser.Services;
 using CefFlashBrowser.Sol;
 using CefFlashBrowser.Utils;
 using SimpleMvvm;
