+++ CefFlashBrowser/Views/PopupWebPage.xaml.cs
@@ -1,6 +1,6 @@
-﻿using CefFlashBrowser.FlashBrowser.Handlers;
 using CefFlashBrowser.Data;
-using CefFlashBrowser.Utils;
+using CefFlashBrowser.FlashBrowser.Handlers;
+using CefFlashBrowser.Services;
 using CefFlashBrowser.WinformCefSharp4WPF;
 using CefSharp;
 using SimpleMvvm.Messaging;
