+++ CefFlashBrowser/Views/SolEditorWindow.xaml.cs
@@ -1,5 +1,5 @@
-﻿using CefFlashBrowser.Models;
-using CefFlashBrowser.Utils;
+using CefFlashBrowser.Models;
+using CefFlashBrowser.Services;
 using CefFlashBrowser.ViewModels;
 using System;
 using System.ComponentModel;
